+++ examples/data-sources/vaultwarden_organization/data-source.tf
@@ -1,0 +1,3 @@
+data "vaultwarden_organization" "example" {
+  id = "53878c48-51e9-416d-b31a-1b4209c93832"
+}